+++ jest.config.js
@@ -4,9 +4,13 @@ module.exports = {
   preset: 'ts-jest',
   roots: ['src/modules'],
   testRegex: '/test/(.+)\\.test\\.ts$',
-  moduleFileExtensions: ["ts", "js", "json"],
+  moduleFileExtensions: ['ts', 'js', 'json'],
   coverageDirectory: '.report',
   coverageReporters: ['lcov', 'html'],
+  setupFiles: [
+    './src/modules/esl-utils/test/deviceDetector.mock.ts',
+    './src/modules/esl-utils/test/matchMedia.mock.ts'
+  ],
   collectCoverageFrom: [
     'src/modules/**/*.ts',
     // cumulative exclude