+++ src/modules/all.ts
@@ -18,6 +18,6 @@ export * from './esl-trigger/core';
 export * from './esl-panel/core';
 
 export * from './esl-tab/core';
-export * from './esl-scrollable-tab/core';
+export * from './esl-scrollable-tabs/core';
 
 export * from './esl-scrollbar/core';