+++ src/modules/esl-utils/async/promise.ts
@@ -1,3 +1,5 @@
+import type {AnyToAnyFnSignature} from '../misc/functions';
+
 /** Interface to describe abstract listenable target */
 export type ListenableTarget = {
   addEventListener: (
@@ -12,11 +14,19 @@ export type ListenableTarget = {
   ) => void;
 };
 
+/** Deferred object represents promise with it's resolve/reject methods */
 export type Deferred<T> = {
+  /** Wrapped promise */
   promise: Promise<T>;
+  /** Function that resolves wrapped promise */
   resolve: (arg: T) => void;
+  /** Function that rejects wrapped promise */
   reject: (arg?: any) => void;
 };
+
+/** Return function type with the same signature but with the result type wrapped into promise */
+export type PromisifyResultFn<F extends AnyToAnyFnSignature> =
+  ((...args: Parameters<F>) => Promise<ReturnType<F> | void>);
 
 /**
  * Promise utils helper class
@@ -43,9 +53,9 @@ export abstract class PromiseUtils {
     options?: boolean | AddEventListenerOptions
   ): Promise<Event> {
     return new Promise((resolve, reject) => {
-      function eventCallback(...args: any) {
+      function eventCallback(e: Event) {
         target.removeEventListener(event, eventCallback, options);
-        resolve(...args);
+        resolve(e);
       }
 
       target.addEventListener(event, eventCallback, options);
@@ -106,7 +116,7 @@ export abstract class PromiseUtils {
   }
 
   /**
-   * Create Deferred Object that wraps promise and it's resolve and reject callbacks
+   * Create Deferred Object that wraps promise and its resolve and reject callbacks
    */
   static deferred<T>(): Deferred<T> {
     let reject: any;