+++ src/modules/esl-utils/misc/functions.ts
@@ -1,10 +1,10 @@
 /**
- * Function that do nothing
+ * Function that does nothing
  */
-export const noop = (...args: any[]): void => undefined;
+export const noop: AnyToVoidFnSignature = () => undefined;
 
 /**
- * Function that return first argument
+ * Function that returns the first argument
  */
 export const identity = <T>(arg: T): T => arg;
 
@@ -14,11 +14,4 @@ export type AnyToVoidFnSignature = (...args: any[]) => void;
 
 export type AnyToAnyFnSignature = (...args: any[]) => any;
 
-/**
- * Find the first defined param
- */
-export function defined<T>(...params: T[]) {
-  for (const param of params) {
-    if (param !== undefined) return param;
-  }
-}
+export type MethodTypedDecorator<T> = (target: any, property: string, descriptor: TypedPropertyDescriptor<T>) => TypedPropertyDescriptor<T> | void;