+++ src/modules/esl-utils/misc/test/object.test.ts
@@ -1,31 +1,95 @@
-import {deepCompare} from '../object';
+import {defined, deepCompare, getPropertyDescriptor} from '../object';
 
-describe('misc/object helper tests', () => {
-  test('deepCompare simple tests', () => {
-    expect(deepCompare({}, {})).toBeTruthy();
-    expect(deepCompare({a: 1}, {a: 1})).toBeTruthy();
-    expect(deepCompare({a: null}, {a: null})).toBeTruthy();
-    expect(deepCompare({a: []}, {a: []})).toBeTruthy();
-    expect(deepCompare({a: 1}, {b: 2})).toBeFalsy();
-    expect(deepCompare({b: 1}, {b: 2})).toBeFalsy();
-    expect(deepCompare({b: undefined}, {b: null})).toBeFalsy();
-    expect(deepCompare({c: NaN}, {c: NaN})).toBeTruthy();
-    expect(deepCompare({a: 1, c: Infinity}, {a: 1, c: Infinity})).toBeTruthy();
+describe('misc/object', () => {
+  describe('deepCompare', () => {
+    test.each([
+      [null, null],
+      [NaN, NaN],
+      ['', ''],
+      [{}, {}],
+      [{a: 1}, {a: 1}],
+      [{a: null}, {a: null}],
+      [{a: []}, {a: []}],
+      [{c: NaN}, {c: NaN}],
+      [{a: 1, c: Infinity}, {a: 1, c: Infinity}],
+      [{a: {b: 1}}, {a: {b: 1}}],
+      [{a: {b: {c: 1}}}, {a: {b: {c: 1}}}],
+      [[], []],
+      [[1], [1]],
+      [[1, 2, 3], [1, 2, 3]],
+      [[{}, {}], [{}, {}]],
+      [[{a: 1}, {b: ''}], [{a: 1}, {b: ''}]]
+    ])('%p should be equal to %p', (a: any, b: any) => expect(deepCompare(a, b)).toBe(true));
+
+    test.each([
+      [undefined, null],
+      [1, 2],
+      ['a', 'b'],
+      [{a: 1}, {b: 2}],
+      [{b: 1}, {b: 2}],
+      [{b: undefined}, {b: null}],
+      [{a: {c: {b: 1}}}, {a: {b: {c: 1}}}],
+      [{a: {a: 1, b: 1}}, {a: {b: 1}}],
+      [{a: {b: 1}}, {a: {a: 1, b: 1}}],
+      [[], [1]],
+      [[1], [1, 2]],
+      [[2, 1], [1, 2]]
+    ])('%p should not be equal to %p', (a: any, b: any) => expect(deepCompare(a, b)).toBe(false));
   });
-  test('deepCompare array tests', () => {
-    expect(deepCompare([], [])).toBeTruthy();
-    expect(deepCompare([1], [1])).toBeTruthy();
-    expect(deepCompare([1, 2, 3], [1, 2, 3])).toBeTruthy();
-    expect(deepCompare([1], [1, 2])).toBeFalsy();
-    expect(deepCompare([2, 1], [1, 2])).toBeFalsy();
-    expect(deepCompare([{}, {}], [{}, {}])).toBeTruthy();
-    expect(deepCompare([{a: 1}, {b: ''}], [{a: 1}, {b: ''}])).toBeTruthy();
+
+  describe('getPropertyDescriptor', () => {
+    test('simple', ()=> {
+      const object = {
+        a: 1,
+        get b() {return 2}
+      };
+      expect(getPropertyDescriptor(object, 'a')).toBeTruthy();
+      expect(getPropertyDescriptor(object, 'b')).toBeTruthy();
+      expect(getPropertyDescriptor(object, 'c')).toBeFalsy();
+
+      const descC = {get: () => 3};
+      const descD = { value: 4 };
+      Object.defineProperty(object, 'c', descC);
+      Object.defineProperty(object, 'd', descD);
+
+      expect(getPropertyDescriptor(object, 'c')).toHaveProperty('get');
+      expect(getPropertyDescriptor(object, 'd')).toHaveProperty('value', 4);
+      expect(getPropertyDescriptor(object, 'f')).toBeFalsy();
+    });
+
+    test('1 lvl', ()=> {
+      class A {
+        a = 1;
+        b() { return 2; }
+        get c() { return 3; }
+      }
+      const obj = new A();
+      expect(getPropertyDescriptor(obj, 'a')).toBeTruthy();
+      expect(getPropertyDescriptor(obj, 'b')).toBeTruthy();
+      expect(getPropertyDescriptor(obj, 'c')).toBeTruthy();
+      expect(getPropertyDescriptor(obj, 'b')).toHaveProperty('value');
+      expect(getPropertyDescriptor(obj, 'c')).toHaveProperty('get')
+    });
+
+    test('2 lvl', ()=> {
+      class A {
+        a() { return 3; }
+        get b() { return 3; }
+      }
+      class B extends A {}
+      const obj = new B();
+      expect(getPropertyDescriptor(obj, 'a')).toBeTruthy();
+      expect(getPropertyDescriptor(obj, 'b')).toBeTruthy();
+    });
   });
-  test('deepCompare deep tests', () => {
-    expect(deepCompare({a: {b: 1}}, {a: {b: 1}})).toBeTruthy();
-    expect(deepCompare({a: {b: {c: 1}}}, {a: {b: {c: 1}}})).toBeTruthy();
-    expect(deepCompare({a: {c: {b: 1}}}, {a: {b: {c: 1}}})).toBeFalsy();
-    expect(deepCompare({a: {a: 1, b: 1}}, {a: {b: 1}})).toBeFalsy();
-    expect(deepCompare({a: {b: 1}}, {a: {a: 1, b: 1}})).toBeFalsy();
+
+  test('defined', () => {
+    expect(defined('a')).toBe('a');
+    expect(defined('', 'a')).toBe('');
+    expect(defined('a', '')).toBe('a');
+    expect(defined(undefined, 'a')).toBe('a');
+    expect(defined(null, 'a')).toBe(null);
+    const obj = {};
+    expect(defined(obj, null)).toBe(obj);
   });
 });